change upload-tartifact to upload-artifact
Workflow was named build-pip-tar because it used upload-tartifact which greatly speeds up the build process since most of the time time is spent uploading the ~10,000+ files. Using tartifact means you get a tarball inside of a zip file when you download the artifact, which is obviously not ideal for users but useful when testing.